requirements.txt
openai==0.27.0
requests==2.26.0
python-dotenv==0.19.1
